refactor : (#68) modify reponse
