+++ src/main/kotlin/entry/dsm/gitauth/equusgithubauth/domain/user/service/UserService.kt
@@ -5,8 +5,8 @@ import entry.dsm.gitauth.equusgithubauth.domain.auth.exception.UnAuthorizedOrgAc
 import entry.dsm.gitauth.equusgithubauth.domain.user.entity.User
 import entry.dsm.gitauth.equusgithubauth.domain.user.entity.enums.UserRole
 import entry.dsm.gitauth.equusgithubauth.domain.user.entity.repository.UserRepository
-import entry.dsm.gitauth.equusgithubauth.domain.user.presentation.dto.response.LoginSuccessResponse
-import entry.dsm.gitauth.equusgithubauth.global.external.github.presentation.controller.GithubApiClient
+import entry.dsm.gitauth.equusgithubauth.domain.user.presentation.dto.response.TokenResponse
+import entry.dsm.gitauth.equusgithubauth.global.infrastructure.feign.client.GithubApiClient
 import entry.dsm.gitauth.equusgithubauth.global.security.jwt.JwtTokenProvider
 import org.springframework.stereotype.Service
 
@@ -21,7 +21,7 @@ class UserService(
         private fun withBearer(accessToken: String) = "Bearer $accessToken"
     }
 
-    fun execute(accessToken: String): LoginSuccessResponse {
+    fun execute(accessToken: String): TokenResponse {
         val bearerToken = withBearer(accessToken)
         val userInfo = githubApiClient.getUser(bearerToken)
         val tokens = jwtTokenProvider.generateToken(userInfo.login)
@@ -44,7 +44,7 @@ class UserService(
             userRepository.save(user)
         }
 
-        return LoginSuccessResponse(
+        return TokenResponse(
             tokens.accessToken,
             tokens.accessTokenExpiration,
             tokens.refreshToken,
